Install the generated .gemspec file
Building the C implementation of facter also produce a .gemspec file.  So far,
it was not installed (probably by mistake: since nothing was using it,
everything was fine).  Now that we are about to add a consumer of this file as
sysutils/rubygem-puppetserver-ca to the ports tree, this file must be
installed.

With hat:	puppet


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@480784 35697150-7ecd-e111-bb59-0022644237b5