add port x11/py-xcffib
xcffib is intended to be a (mostly) drop-in replacement for xpyb. xpyb has
an inactive upstream, several memory leaks, is python2 only and doesn't have
pypy support. xcffib is a binding which uses cffi, which mitigates some of
the issues described above. xcffib also builds bindings for 27 of the 29
(xprint and xkb are missing) X extensions in 1.10.

Approved by:	wg (mentor)