MFH: r481915 r481917
sysutils/ipmitool:  unbreak with openssl 1.1.1 import
- Only apply openssl patch if we are on a version of FreeBSD with openssl 1.1.1
- Don't bump portrevision as we don't change anything except on broken systems

It should be noted that this is a functional way to fix this port and is the method
used upstream.  There are most likely better ways to do this.

Reviewed by:    0mp (Makefile changes)
Sponsored by:   Limelight Networks

Setting DOCS=off will set EXTRA_PATCHES, so we need to append to this
variable, not clear it.

Submitted by:	tobik