net/mpd5: make it possible to build with ssl=openssl111
Now as FreeBSD 12 got openssl-1.1.1 in base system and
/usr/include/openssl/ssl.h and /usr/local/include/openssl/ssl.h
are indentical, it is safe to allow building net/mpd5
with DEFAULT_VERSIONS having ssl=openssl111 and no reasons to break
the build.

However, resulting binary still links with base system /usr/lib/libssl.so
because of libfetch(3) needs it and mpd5 links with libfetch too.