graphics/freeglut: Fix build after joy(4) removal
Fix build of graphics/freeglut after the removal of the joy(4) driver in
FreeBSD CURRENT.  The same fix is applied for all FreeBSD versions though.
Simply remove the include of sys/joystick.h, and instead bring in the
definition of struct joystick from that file into the freeglut source.  None
of the ioctls in sys/joystick.h are ever used in freeglut.

Please note that juystick support in freeglut on FreeBSD is not really
tested, so no promises are made about it.

Reported by:	goshanecr on the FreeBSD forum
in collaboration with:	imp