Add all patches from base llvm/clang/lld/lldb 7.0 to devel/llvm70
This adds all the patches that were applied in the past to the
clang700-import branch, under contrib/llvm. After these, there only
minimal diffs left between the port sources and the base sources.

Most of these remaining diffs are due to #ifdef shortcuts in the base
sources, because we don't compile certain features in. Other diffs are
because the port has applied a few changes that we don't have in base.

Also switch to the common LICENSE defintion in devel/llvm-devel and
chase new USE_GNOME requirements (for libxml2).

PR:	212343, 230604
Submitted by:	dim
MFH:		2018Q4
Differential Revision:	https://reviews.freebsd.org/D17709


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@483690 35697150-7ecd-e111-bb59-0022644237b5