databases/percona57-server: Adapt ${name}_limits to the new resource limiting mechanics for services
Thanks to the recent change to rc.subr(8) it is no longer necessary to
define a resource limiting logic in service scripts. limits(1) is now run
by default and its configuration is handled via the "${name}_limits"
variable.

Unfortunately, there is a collision of variable names. The service scripts
are broken on FreeBSD version with that new change (>= 12.0).

The patch preserves the old behavior on the systems without the rc.subr(8)
modification.

PR:		227236
Reviewed by:	krion
Approved by:	krion (mentor), maintainer timeout
Differential Revision:	https://reviews.freebsd.org/D17786