Align *_DEPENDS and use paths for bash in my ports; fix var name.
On systems where bash is in $PATH but not in ${LOCALBASE}/bin/bash,
ilmbase would fail to build, and e2fsprogs might fail if the
BASHTESTS option was enabled.
Make sure that the _DEPENDS and the actual path handed to the configure or
test rig match to the letter, to avoid a time-of-check vs. time-of-use
discrepancy.

In e2fsprogs, the helper variable was also mistyped and would never
amend to BUILD_DEPENDS. Fix typo.

Check my other ports, but wiggle and xmlto [textproc/] seem unaffected.

Reported by:	J R Matthews (direct email)
MFH:		2018Q4


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@486634 35697150-7ecd-e111-bb59-0022644237b5