Once more into the breach for net/qt4-network.
This time it was build and run-tested in six configuration variants of 11.2
(all the valid settings for ssl= in DEFAULT_VERSIONS). Not tested by me in
12 though, others say it does work.

With libressl it still doesn't *function*, but it builds. As far as I'm
concerned for EOL software, that's so far outside the support envelope
as to be ignored.

PR:		214691
Submitted by:	jkim
Obtained from:	Debian