Mark broken with unsupported OpenSSL release
puppetdb-cli internal dependencies support OpenSSL 1.0 or older only. To
sum up, one will be able to build/run puppetdb-cli using SSL from base
(FreeBSD <= 11) or from security/openssl (all FreeBSD versions).

Mark broken for:
  - DEFAULT_VERSIONS+=ssl=base (for FreeBSD 12+)
  - DEFAULT_VERSIONS+=ssl=openssl111
  - DEFAULT_VERSIONS+=ssl=libressl

An experimental patch to bring support for OpenSSL 1.1.1 is available in
the PR 233389.
The details why we chose not to merge it is explained in the comments.

PR:             233389
Reported by:    jbeich
With hat:       puppet