Update to new upstream release 5.5.
Credits to Greg for a thorough preparation of the update,
much appreciated with a huge thank you.

Release announcement:
<https://rawtherapee.com/blog/rawtherapee-5.5-released>

A selection of news (read the announcement for more, and screenshots):
* stripe artifact removal for phase detection auto-focus (PDAF)
  for Nikon Z6, Z7, and various Sony cameras
* improved support for Canon mRaw
* new Shadows/Highlights tool
* constrast threshold masks to filter effects by high/low detail regions
* new color toning methods
* new soft light tool
* ICC profile creator
* new "flexible" tone curve type
* improved lensfun chromatic aberration correction

While here, since the abort-on-exit bug persists,
wrap rawtherapee in a script that runs ulimit -c 0,
and make note of it in pkg-message. References:
* <https://bugs.freebsd.org/bugzilla/show_bug.cgi?id=205370>
* <https://github.com/Beep6581/RawTherapee/issues/3636>

PR:		234153
Submitted by:	Greg V <greg@unrelenting.technology>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@488088 35697150-7ecd-e111-bb59-0022644237b5