- Fix build with SSL on 12.0
Vsftpd uses an ugly way of checking whether it needs to link
libssl/libcrypto by grepping object file (!) for SSL_library_init
(which may be or not be actually used depending on a define).
This does not work if SSL_library_init is a macro which is the case
with OpenSSL 1.1.0:

/usr/include/openssl/ssl.h:#  define SSL_library_init() OPENSSL_init_ssl(0, NULL)

Because of this, on 12.0 the check always fails, OpenSSL libs are
not linked, and the build fails with linker errors:

/usr/bin/ld: error: undefined symbol: OPENSSL_init_ssl
>>> referenced by ssl.c
>>>               ssl.o:(ssl_init)

/usr/bin/ld: error: undefined symbol: TLS_server_method
>>> referenced by ssl.c
>>>               ssl.o:(ssl_init)

/usr/bin/ld: error: undefined symbol: SSL_CTX_new
>>> referenced by ssl.c
>>>               ssl.o:(ssl_init)

Since we know beforehand whether we need SSL, just replace the check
with true if SSL is enabled.

- While here, partially convert to options helpers