Update audio/rosegarden to 18.12 (latest release).
Release Notes since 17.04 (previous version in ports):
	https://rosegardenmusic.com/wiki/dev:17.12
	https://rosegardenmusic.com/wiki/dev:18.06
	https://rosegardenmusic.com/wiki/dev:18.12

Notable is removing Qt4 support, which aligns well with our intentions
to remove Qt4 from ports. Since Qt5 was the default toolkit already,
this should have little impact on Rosegarden users.

Each of the releases had a bunch of bugfixes as well.

This update supercedes the update in the PR, which bumped to version 17.12.

PR:		227384
Reported by:	yuri


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@488721 35697150-7ecd-e111-bb59-0022644237b5