Update audio/musescore to 3.0.0, the latest release
This skips over 18 months of musescore development to get to
what's new. This version uses WebEngine, although I don't know
what it uses it for; a PORTREVISION should be able to clear that
up and make it available for aarch64 and ppc64.

All the local patches have been dropped, except for some extra
CMake-futzing in order to get it to build with ninja and not
duplicate whole chunks of WebEngine.

PR:		234801
Reported by:	kunda


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@489895 35697150-7ecd-e111-bb59-0022644237b5