Add p5-Data-Binary 0.01
Data::Binary provides string equivalents to the -T / -B operators. Since these
only work on file names and file handles, this module provides the same
functions but on strings.

Note that the actual implementation is currently different, basically because
the -T / -B functions are in C/XS, and this module is written in pure Perl. For
now, anyway.

WWW: https://metacpan.org/release/Data-Binary