security/sandsifter: update to version 1.03
This version is based on a fork of the original repository (which was
itself a fork) and is the same fork that Debian and Ubuntu use. The
previous repository is no longer under development.

Some binaries have been renamed by the new upstream.

Extend pkg-messsage with some more usage instructions.

Reported by:	repology


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@490792 35697150-7ecd-e111-bb59-0022644237b5