x11/libxkbcommon: Add dep on xkeyboard-config
Add a run time dependency on x11/xkeyboard-config in x11/libxkbcommon.
While not strictly necessary, this is recommended upstream, since almost all
uses of libxkbcommon also needs xkeyboard-config.  This is similar to how it
is done in other distributions.

This solves issues when graphics/xpdf4 is installed without a desktop or
graphical environment, as well as other places.

PR:		229127
Reported by:	Bertrand Petit
Suggested by:	tobik