Add a new atf-master master port. ARM Arm Trusted Framework (sometimes called TF-A for reasons ...) provides a reference implementation of secure world software for Armv7-A and Armv8-A, including a Secure Monitor executing at Exception Level 3 (EL3).
We had a port for Alwinner A64/H5 chip, but we need to introduced more ports.
So convert this to a master/slave port scheme like u-boot.

atf-allwinner is now badly named, when it was introduced it only support A64,
now it both support A64 and H5 and there is another ATF platform for H6.
Rename it with the platform name