New port: x11-fonts/3270font: IBM 3270 font for the nostalgic
This font is derived from the x3270 font, which, in turn, was translated
from the one in Georgia Tech's 3270tool, which was itself hand-copied
from a 3270 series terminal. This font was built because the author felt
that terminals deserve to be pretty. The .sfd font file contains a x3270
bitmap font that was used for guidance.

WWW: https://github.com/rbanffy/3270font

Reviewed by:	mat
Approved by:	mat (mentor)
Differential Revision:	https://reviews.freebsd.org/D19186


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@492915 35697150-7ecd-e111-bb59-0022644237b5