devel/erlang-exmpp: Use gsed during the build
Parts of erlang-exmpp's build use base sed with expressions that
don't actually work correctly in our sed as of yet.  In this case,
the replacement of \xa9 with a &copy; entity.  The former should
be interpreted as the character \xa9 (copyright symbol), but sed
does not currently interpret it as such (or \r, \n, \t).  Switch
to textproc/gsed for now, which does the right thing.

PR:		233424
Submitted by:	kevans
Approved by:	erlang (maintainer timeout, 3 months)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@493950 35697150-7ecd-e111-bb59-0022644237b5