net-mgmt/nagiosql: Remove redundant regex escapes
nagiosql's Makefile currently has some redundant escaping of single
quotes in its Makefile.  These get passed through as escaped single
quotes to regex(3), which is an escape of an ordinary character.
Such escapes will be forbidden in a future version of regex(3), so
we'll get ahead of the game and remove said escapes now.

While here, there are two questionable expressions: -e
"s|\>/etc/nagiosql|\>${ETCDIR}|g" and the one immediately following
it.  \> is a word delimiter (see reformat(7)), and will only match
as such.  If this were to match, it would replace <word
delimiter>/etc/nagiosql with >${ETCDIR} -- this likely works as
intended now, but I think it was intended that it matches a literal
>/etc/nagiosql, so the included patch treats it accordingly.

- While here reset MAINTAINER after a timeout of 3 months and
  repeated previous timeouts

PR:		233458
Submitted by:	kevans
Approved by:	chris@smartt.com (maintainer timeout, 3 months)