With libc++ 8.0, which is in the projects/clang800-import branch, and which will soon be merged to head, compilation of devel/jsoncpp fails due to a conflict between the new C++ <version> header, and a local file "version" which is produced by jsoncpp during its configure phase.
This is due to the initial test runner being compiled with "-I.", even
though it does not need any file from the port's working directory root.
Therefore, it seems to be easiest to comment out the line from the
SConstruct script that adds the "-I." option.

Approved by:	yuri (maintainer)
PR:		236061
MFH:		2019Q1