[NEW PORT] net/measurement-kit: Portable C++14 network measurement library
Measurement Kit is a C++14 library that implements open network measurement
methodologies (performance, censorship, etc.) on Unix-like operating systems
such as Android, iOS, Linux, and macOS.

It is meant to be embedded by third party applications with specific network
measurement needs and/or to be used by researchers to implement novel
network measurement tools. It started as a common engine to implement the
mobile versions of OONI and Neubot, and since then has grown to include also
the implementation of Portolan traceroute and NDT tests.

  WWW: https://measurement-kit.github.io/

Approved by:	rene (mentor)
Differential Revision:	https://reviews.freebsd.org/D19342


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@494800 35697150-7ecd-e111-bb59-0022644237b5