New port: x11/py-caffeine-ng.
Caffeine is a little daemon that sits in your systray, and prevents
the screensaver from showing up, or the system from going to sleep. It
does so when an application is fullscreened (eg: youtube), or when you
click on the systray icon (which you can do, when, eg: reading).

WWW: https://github.com/caffeine-ng/caffeine-ng

PR:		235639
Submitted by: 	Hiroki Tagato <tagattie@yandex.com>