Turn on RAR option by default only on x86 in archivers/ark.
This makes it possible to build on non-x86 with default options. Support
for rar archives is via external executables, so there is no change in the
port's compiled artifacts.

Thanks linimon@ for the functional part of the change. While here,
polish up descriptions and links and bits and bobs.

PR:		236240
Submitted by:	linimon