www/node10: set environment variables for host build
The Node.js build system supports cross-compilation via node-gyp, which
picks up CC, CXX, LINK, C*FLAGS, and LDFLAGS from the environment for the
target build, but not for the host. We need to set the environment variables
for the "host" build.

This unbreaks the build on powerpc64.

PR:		233835
Reported by:	Piotr Kubaj <pkubaj@anongoth.pl>
Sponsored by:	Miles AS
Differential Revision:	https://reviews.freebsd.org/D19250