emulators/qemu-sbruno:  - Add a "-O argv0" argument to qemu-bsd-user for use in execve() calls:
Prior to this commit, the value of argv[0] was discarded as it was
overwritten by the path to the program being executed.  This commit
uses the new -0 option to retain that information; in particular,
this unbreaks mailwrapper+sendmail, which makes use of this to know
which of multiple functions should be performed.

Submitted by:	cperciva