Remove neovim from TEST_DEPENDS, and original submitter becomes maintainer
My intention is to add an OPTION for pynvim to the neovim port, which
means that this port cannot depend upon neovim. Change do-test to a
maintainer-test target that gives some instructions.

Also, per original PR and at request of 0mp, pass maintainership to
Thomas Merkel, the submitter of the original py-pynvim PR.

PR:		236571
Approved by:	0mp