Add a port of django-extensions specifically for Django 2.1
This uses the approach taken by kai@ for various django ports
in r492690: Depend on Django 2.1, require python 3.5 or
higher and add a conflict variable to the Django 1.x version
of the port.

This also updates the port to 2.1.6, while the Django 1.x
version remains at 2.1.5 for the time being.

PR:		236682