net/tigervnc: remove unnecessary args from CMAKE_ARGS
as this argument is not necessary. ${STAGE} was typo of ${STAGEDIR}
however, ${STAGEDIR} should not be put here. ${STAGEDIR}${PREFIX} is not
correct. Fortunately, because of this typo, it was working correctly.

> -DCMAKE_INSTALL_PREFIX=${STAGE}${PREFIX}

Furthermore, the following argument is set by Mk/bsd.cmake.mk.

> -DCMAKE_INSTALL_PREFIX:PATH="${PREFIX}"

Both arguments are entirely same and redundant so removing this is definitely
safe and no-op.

Sponsored by:	HAW International, Inc.