www/qt5-webengine: Update to 5.12.1
It was again a long journey to bring the port in sync with the other Qt5
ports. It runs now under the hood with Chromium 69.0.3497.128 and applied
security fixes up to version 71.0.3578.94.

Also the approach with the handling of the BUILD.gn files has changed. With
the previous version of www/qt5-webengine (and www/chromium up to r449991) the
boolean variables "is_bsd" and "is_posix" were set to true in BUILDCONFIG
when FreeBSD was detected as operating system during the build process.

Now the boolean variable "is_linux" is set to true as well and this reduces
some patching of BUILD.gn files. It makes it (hopefully) also somewhat easier
to exclude Linux-only features from future www/qt5-webengine versions.

The .debug entries in the pkg-plist are prefixed with @comment each as a
temporary workaround because building with "separate_debug_info" fails to
build QtWebEngineProcess at the moment. (see also: QTBUG-74312)

Many kudos must go to the FreeBSD Chromium team for their ongoing efforts to
keep the browser in a good shape. I was able to use a decent amount of
patches from there which sped up the whole process.

At last but not least: Many thanks must also go to rakuco@ for finding the
last pieces to make www/qt5-webengine more usable and tcberner@ for the moral
support during the whole time.

Changelogs:

https://code.qt.io/cgit/qt/qtwebengine.git/tree/dist/changes-5.12.0
https://code.qt.io/cgit/qt/qtwebengine.git/tree/dist/changes-5.12.1

PR:		234470 [1] (related), 235075 [2]
Reported by:	Michael Danilov [2], aeuii@posteo.de [1]
Reviewed by:	rakuco, tcberner (mentor)
Approved by:	rakuco, tcberner (mentor), kde (maintainer)
Differential Revision:	https://reviews.freebsd.org/D18757


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@496989 35697150-7ecd-e111-bb59-0022644237b5