blind-fix build of devel/pyside2
pyside2 fails to build on the package cluster, e.g.
http://beefy5.nyi.freebsd.org/data/120i386-default/498628/logs/py27-pyside2-5.12.2.log
I cannot reproduce that failure locally, but I sincerly hope that the
additional patch fixes it. Add a comment, else I'll forget that patch
is required.
While here, update the pre-existing patch file to reduce confusion when
looking at patch and code (patch was fine with the file, which did surprise
me).