Re-add x11/qimageblitz: Graphical effects and filters library, part of kdesupport
The older version of this port has been recently removed because it was Qt4-only.
As it turned out, it is used by some software packages, and the latest revision
supports Qt5.

It is also a part of the collection 'kdesupport', see https://launchpad.net/kdesupport
Followed the advice of Adriaan de Groot <groot@kde.org> to use PORTVERSION=0.1.0