lang/erlang: split man pages into lang/erlang-man, and add the FOP/PDF options.
Erlang man pages are delivered as a separate distfile; with this commit
lang/erlang depends on the GitHub sources only. The MANPAGES option now just
adds a dependency on lang/erlang-man.

This allows man pages to be upgraded independently and cover more recent
Erlang runtimes.

The FOP option has been added to allow building the full documentation from
sources. When enabled, man pages will be built and installed based on the
latest sources. When PDF is enabled, documentation in PDF format will also be
installed. These options are disabled by default as they depend on
textprop/fop, and thus Java/X11.

On Erlang 21 the Corba applications have been moved to a separate repository
upstream, but their build still requires a full source tree, so they are not
really independent.

FOP/PDF are provided for when Corba support will be merged back after the
Erlang 21 upgrade. This will make it possible to build the documentation that
has been removed upstream and is no longer delivered in the prebuilt
documentation distfile.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@498799 35697150-7ecd-e111-bb59-0022644237b5