astro/wmsun: "Update" to 1.05 and return to official upstream repository
While I'm here:
* Add license information
* Sort variables a bit according to the PHB

This was initially a small patch that was meant to update the MASTER_SITES
and the WWW field. [1] On further research it came to light that the
previous version of astro/wmsun was a fork whose upstream sources are now
gone.

In the meantime there have been some recent activities regarding Window
Maker dockapps and many of those programs were collected, updated and are
maintained by the current Window Maker developers.

Therefore the intial patch has been extended a bit to point to the new
upstream location instead of using a semi-official location that could
vanish in near future. [2]

At last but not least: The program was renamed from "wmSun" to "wmsun" so
there's a bit more churn than usual.

https://github.com/window-maker/dockapps/blob/master/wmsun/ChangeLog

PR:		235526
Submitted by:	Marco Beishuzen (maintainer) [1]
Approved by:	tcberner (mentor), maintainer [2]
Differential Revision:	https://reviews.freebsd.org/D19136