Update net-im/libqmatrixclient and consumers
New release of libqmatrixclient
	https://github.com/QMatrixClient/libqmatrixclient/releases/tag/0.5.1.2
with crash fixes when there are expired tokens or abandoned connections.
This helps when clients are stuck showing some dialog and the connection
goes away under water.

New RC of quaternion, using the updated library.
	https://github.com/QMatrixClient/Quaternion/releases/tag/0.0.9.4-rc3
Needs a PORTREVISION bump because I'm using DISTVERSIONSUFFIX. This
will go away when 0.0.9.5 is reached.

Bump spectral so that it picks up the new libqmatrixclient.