Upgrade to version 1.2.8.
This package exposes a bug (IMHO) in "pkg" which deletes the directory
pointed to by a symbolic link after removing the files within, but is
not able to recreate that directory when a package is installed.

It may be necessary to manually recreate ${LOCALBASE}/share/nls/C after
removal of this package, e.g. during a port upgrade.

This commit is meant to provide a test case for package maintainers
that hopefully will lead to this issue being fixed in "pkg".

Approved by:	antoine (implicit)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@499433 35697150-7ecd-e111-bb59-0022644237b5