security/p5-Net-SSLeay: fix build with libressl 2.9
I'm considering a maintainer timeout for this one, as the bug was first
reported in December. LibreSSL 2.9.1 is now the main version in the ports tree,
so this is urgent.

perl@, feel free to commit a better fix.

PR:		234102
Submitted by:	Charlie Li <ml+freebsd@vishwin.info>
Approved by:	maintainer (timeout)