New port: net/py-aiocoap -- The Python CoAP library
The aiocoap package is an implementation of CoAP, the Constrained Application
Protocol.

It is written in Python 3 using its native asyncio methods to facilitate
concurrent operations while maintaining an easy to use interface.

aiocoap is originally based on txThings. If you want to use CoAP in your
existing Twisted application, or can not migrate to Python 3 yet, that is
probably more useful to you than aiocoap.

WWW: https://github.com/chrysn/aiocoap


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@500359 35697150-7ecd-e111-bb59-0022644237b5