New port: textproc/markdownfmt: Tool for linting and formatting Markdown in-place (similar to gofmt)
markdownfmt is a linter similar to gofmt but for the Markdown language.
It is capable of autoformatting Markdown files in-place or generating diffs
with linting applied. Also, it is used in many plugins for formatting
Markdown files written for text editors like Vim or Emacs.

WWW: https://github.com/shurcooL/markdownfmt


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@501698 35697150-7ecd-e111-bb59-0022644237b5