Update amazon-ssm-agent to 2.3.612.0.
This adds support for "sessions", aka. creating a remote shell via the Agent.

In order to support this, an "ssm-user" account is created when the package
is installed; we create this via pkg-install rather than the USERS= / UIDs
functionality, since ssm-user needs to be in wheel and needs to have a home
directory.

Thanks to:	Alexey Milevsky


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@501759 35697150-7ecd-e111-bb59-0022644237b5