Fix named when using plugins and chroot.
BIND9 introduced plugins and migrated the filter-aaaa feature to a
plugin.
As it loads its plugins late in the startup process (read after chroot),
the plugins need to be available in the chroot.

Also, refactor the code now that a second directory need to be handled.

PR:		238011
Reported by:	ryan@timewasted.me
MFH:		2019Q2


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@503955 35697150-7ecd-e111-bb59-0022644237b5