Add py-gym 0.12.5
OpenAI Gym is a toolkit for developing and comparing reinforcement learning
algorithms. This is the gym open-source library, which gives you access to a
standardized set of environments.

gym makes no assumptions about the structure of your agent, and is compatible
with any numerical computation library, such as TensorFlow or Theano. You can
use it from Python code, and soon from other languages.

There are two basic concepts in reinforcement learning: the environment (namely,
the outside world) and the agent (namely, the algorithm you are writing). The
agent sends actions to the environment, and the environment replies with
observations and rewards (that is, a score).

The core gym interface is Env, which is the unified environment interface. There
is no interface for agents; that part is left to you. The following are the Env
methods you should know:
- reset(self): Reset the environment's state. Returns observation.
- step(self, action): Step the environment by one timestep. Returns observation,
  reward, done, info.
- render(self, mode='human'): Render one frame of the environment. The default
  mode will do something human friendly, such as pop up a window.

WWW: https://gym.openai.com/
WWW: https://github.com/openai/gym


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@504818 35697150-7ecd-e111-bb59-0022644237b5