biology/viennarna: Make it less/not prone to failures due to hardcoded C++ library name
configure and Makefile.in provided in tarballs have -lstdc++ hard-coded in them.
The previous version of the port patches these to be -lc++, which still might fail on gcc systems, depending on which C++ library is used.

This change adds USES=autoreconf which makes it to regenerate configure and Makefile.in and hopefully have the proper C++ library in these files.

GNU libtool links with the C compiler, not with the C++ compiler, in this project. This is what I think triggers link failures due to a wrong C++
library getting into the link lines. GNU libtool likely chooses C compiler because this project has its 'main' functions in .c files,
but also includes C++ object modules. This likely causes GNU tools confusion and it uses the C compiler to link and fails due to wrong C++ libraries added there.