Update KDE Plasma Desktop to 5.16.2
Tuesday, 25 June 2019. Today KDE releases a Bugfix update to KDE Plasma 5,
versioned 5.16.2. Plasma 5.16 was released in June with many feature
refinements and new modules to complete the desktop experience.

This release adds a week's worth of new translations and fixes from KDE's
contributors. The bugfixes are typically small but important and include:
 * Klipper Always restore the last clipbboard item. Commit.
 * Discover: Improved notification identification for Snap applications.
 * Notifications: Don't keep non-configurable notifications in history.

Announcement:
	https://kde.org/announcements/plasma-5.16.2.php

The changelog can be found here:
	https://kde.org/announcements/plasma-5.16.1-5.16.2-changelog.php


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@505128 35697150-7ecd-e111-bb59-0022644237b5