Add p5-JSON-Parse 0.55
JSON::Parse is a module for parsing JSON. (JSON means "JavaScript Object
Notation" and it is specified in "RFC 7159".)

JSON::Parse offers the function "parse_json", which takes a string containing
JSON, and returns an equivalent Perl structure. It also offers validation of
JSON via "valid_json", which returns true or false depending on whether the JSON
is correct or not, and "assert_valid_json", which produces a descriptive fatal
error if the JSON is invalid. A function "json_file_to_perl" reads JSON from a
file, and there is a safer version of "parse_json" called "parse_json_safe"
which doesn't throw exceptions.

For special cases of parsing, there are also methods "new" and "run", which
create a JSON parsing object and run it on text. See "METHODS".

JSON::Parse accepts only UTF-8 as input. See "UTF-8 only" and "Handling of
Unicode".

WWW: https://metacpan.org/release/JSON-Parse