Fix devel/ddd after fallout from libXt update.
Stricter type rules mean that you can't pass a non-noreturn function
as a parameter to a function expecting a noreturn-function. Error
reported by Antoine was

exit.C:815:12: error: no matching function for call to 'XtAppSetErrorHandler'
    (void) XtAppSetErrorHandler(app_context, ddd_xt_error);

because ddd_xt_error had the wrong attribute (non-noreturn). So add
that in, and add it to goto_main_loop() as well, where it makes sense
because that uses longjmp() and doesn't return.

Reported by:	antoine
Reviewed by:	zeising